pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mas-realm-bench"
version = "0.1.0"
description = "Multi-Agent System with RAG evaluation on REALM-Bench"
authors = ["Jan <jan@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
# Core frameworks
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"

# Model providers
openai = "^1.40.0"
anthropic = "^0.34.0"
groq = "^0.9.0"

# Code execution and tools
jupyter-client = "^8.6.0"
ipykernel = "^6.29.0"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
numpy = "^1.24.0"

# RAG and retrieval (optional)
pinecone-client = {version = "^4.1.0", optional = true}
chromadb = {version = "^0.4.0", optional = true}
sentence-transformers = {version = "^2.2.0", optional = true}

# Configuration and utilities
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
rich = "^13.7.0"
asyncio-mqtt = "^0.16.0"

# Testing and development
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.poetry.extras]
rag = ["pinecone-client", "chromadb", "sentence-transformers"]

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipython = "^8.18.0"

[tool.poetry.scripts]
mas-run = "scripts.run_experiments:main"
mas-eval = "scripts.evaluate_metrics:main"

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "W", "C", "N", "UP", "S", "B", "A", "SIM", "I"]
ignore = ["E501", "S101", "S311"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"